Updated myjobs setup and got mypartners test_views working.
