+++ myjobs/tests/test_views.py
@@ -1,119 +1,34 @@
-import base64
-import json
-import time
-import uuid
-from bs4 import BeautifulSoup
 from datetime import timedelta, date
-from importlib import import_module
+import time
+import json
+import uuid
 from urllib import urlencode
 
+import base64
+from bs4 import BeautifulSoup
 from django.conf import settings
-from django.contrib.auth import login
 from django.contrib.sessions.models import Session
 from django.core import mail
 from django.core.urlresolvers import reverse
-from django.http import HttpRequest
 from django.template import Context, Template
-from django.test.client import Client, MULTIPART_CONTENT
-from mymessages.models import Message
-from mymessages.tests.factories import MessageInfoFactory
+from jira.client import JIRA
 
-from setup import MyJobsBase
 from myjobs.models import User, EmailLog, FAQ
 from myjobs.tests.factories import (UserFactory, RoleFactory, ActivityFactory,
                                     AppAccessFactory)
+from myjobs.tests.setup import MyJobsBase, TestClient
+from mymessages.models import Message
+from mymessages.tests.factories import MessageInfoFactory
 from mypartners.tests.factories import PartnerFactory
+from myprofile.models import Name, Education
 from myprofile.tests.factories import SecondaryEmailFactory
 from mysearches.models import PartnerSavedSearch
-from seo.tests.factories import CompanyFactory, CompanyUserFactory
-from myprofile.models import Name, Education
 from mysearches.models import SavedSearch, SavedSearchLog
-from registration.models import ActivationProfile
 from registration import signals as custom_signals
-
+from registration.models import ActivationProfile
 from secrets import options, my_agent_auth
-from jira.client import JIRA
-
+from seo.tests.factories import CompanyFactory, CompanyUserFactory
 from tasks import process_batch_events
-
-
-class TestClient(Client):
-    """
-    Custom test client that decouples testing from the authentication bits, as
-    well as reduces boilerplate when sending requests.
-    """
-
-    def __init__(self, enforce_csrf_checks=False, path=None,
-                 data=None, **defaults):
-        """
-        In addition to Django's test client, this method also takes an optional
-        path and data attribute to be used for get and post requests.
-        """
-        self.path = path
-        self.data = data or {}
-        super(TestClient, self).__init__(enforce_csrf_checks, **defaults)
-
-    def get(self, path=None, data=None, follow=False, secure=False, **extra):
-        """
-        Like the builtin get method, but uses the instances path and data when
-        available.
-        """
-        path = path or self.path
-
-        if not path:
-            raise TypeError("get expects a path. None given. Either "
-                            "instantiate TestClient with a default path or be "
-                            "sure that the first argument to get is a valid "
-                            "path.")
-        data = data or self.data
-
-        return super(TestClient, self).get(
-            path, data=data, follow=follow, secure=secure, **extra)
-
-    def post(self, path=None, data=None, content_type=MULTIPART_CONTENT,
-             secure=False, **extra):
-        path = path or self.path
-        data = data or self.data
-
-        try:
-            return super(TestClient, self).post(
-                path, data=data, content_type=content_type,
-                secure=secure, **extra)
-        except TypeError:
-            raise Exception("Calls to TestClient's methods require that "
-                            "either path be passed explicit, or the "
-                            "path be specified in the constructor")
-
-    def login_user(self, user):
-        if 'django.contrib.sessions' not in settings.INSTALLED_APPS:
-            raise AssertionError("Unable to login without "
-                                 "django.contrib.sessions in INSTALLED_APPS")
-        user.backend = "%s.%s" % ("django.contrib.auth.backends",
-                                  "ModelBackend")
-        engine = import_module(settings.SESSION_ENGINE)
-
-        # Create a fake request to store login details.
-        request = HttpRequest()
-        if self.session:
-            request.session = self.session
-        else:
-            request.session = engine.SessionStore()
-        login(request, user)
-
-        # Set the cookie to represent the session.
-        session_cookie = settings.SESSION_COOKIE_NAME
-        self.cookies[session_cookie] = request.session.session_key
-        cookie_data = {
-            'max-age': None,
-            'path': '/',
-            'domain': settings.SESSION_COOKIE_DOMAIN,
-            'secure': settings.SESSION_COOKIE_SECURE or None,
-            'expires': None,
-        }
-        self.cookies[session_cookie].update(cookie_data)
-
-        # Save the session values.
-        request.session.save()
 
 
 class MyJobsViewsTests(MyJobsBase):
